Refactor import statements hook, manifest models and enums for domain model
