+++ apps/api/src/manifest/models/index.ts
@@ -1,1 +1,2 @@
 export * from './manifest-service.model';
+export * from './manifest-request.model';
