+++ apps/api/src/models/fields/index.ts
@@ -1,2 +1,6 @@
 export * from './field.model';
 export * from './validator.model';
+export * from './const.model';
+export * from './ui.model';
+export * from './on-match-type.enum';
+export * from './validator-type.enum';
