+++ apps/api/src/models/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './hook.model';
+export * from './hook-type.enum';
